Splits HAL Digital Implemntation files
Split to match the new headers. Uses a namespace 'hal' for internal functions and globals.
SPIAccumulator merged back into SPI header, as it was not a good split. Will move analog
accumulator back to analog input when I do the analog split.
