+++ hal/include/HAL/DIO.h
@@ -10,6 +10,8 @@
 #include <stdint.h>
 
 extern "C" {
+// the following 2 functions are here as they will be changed with
+// the handle changes to be DIO exclusive.
 void* initializeDigitalPort(void* port_pointer, int32_t* status);
 void freeDigitalPort(void* digital_port_pointer);
