+++ wpilibj/src/athena/cpp/lib/SPIJNI.cpp
@@ -12,7 +12,6 @@
 #include "edu_wpi_first_wpilibj_hal_SPIJNI.h"
 
 #include "HAL/SPI.h"
-#include "HAL/SPIAccumulator.h"
 #include "HALUtil.h"
 
 // set the logging level
